Add command overload builders
This should make command builders a lot cleaner for lib users and
also makes #264 possible.
