+++ src/Discord.Net.Commands/Builders/CommandBuilder.cs
@@ -7,41 +7,35 @@ namespace Discord.Commands.Builders
 {
     public class CommandBuilder
     {
-        private readonly List<PreconditionAttribute> _preconditions;
-        private readonly List<ParameterBuilder> _parameters;
         private readonly List<string> _aliases;
+        private readonly List<OverloadBuilder> _overloads;
 
         public ModuleBuilder Module { get; }
-        internal Func<CommandContext, object[], IDependencyMap, Task> Callback { get; set; }
 
         public string Name { get; set; }
         public string Summary { get; set; }
         public string Remarks { get; set; }
-        public RunMode RunMode { get; set; }
-        public int Priority { get; set; }
 
-        public IReadOnlyList<PreconditionAttribute> Preconditions => _preconditions;
-        public IReadOnlyList<ParameterBuilder> Parameters => _parameters;
         public IReadOnlyList<string> Aliases => _aliases;
+        public IReadOnlyList<OverloadBuilder> Overloads => _overloads;
 
         //Automatic
         internal CommandBuilder(ModuleBuilder module)
         {
             Module = module;
 
-            _preconditions = new List<PreconditionAttribute>();
-            _parameters = new List<ParameterBuilder>();
             _aliases = new List<string>();
+            _overloads = new List<OverloadBuilder>();
         }
         //User-defined
-        internal CommandBuilder(ModuleBuilder module, string primaryAlias, Func<CommandContext, object[], IDependencyMap, Task> callback)
+        internal CommandBuilder(ModuleBuilder module, string primaryAlias, Action<OverloadBuilder> defaultOverloadBuilder)
             : this(module)
         {
-            Discord.Preconditions.NotNull(primaryAlias, nameof(primaryAlias));
-            Discord.Preconditions.NotNull(callback, nameof(callback));
-
-            Callback = callback;
+            Preconditions.NotNull(primaryAlias, nameof(primaryAlias));
+            Preconditions.NotNull(defaultOverloadBuilder, nameof(defaultOverloadBuilder));
+            
             _aliases.Add(primaryAlias);
+            AddOverload(defaultOverloadBuilder);
         }
 
         public CommandBuilder WithName(string name)
@@ -59,39 +53,17 @@ namespace Discord.Commands.Builders
             Remarks = remarks;
             return this;
         }
-        public CommandBuilder WithRunMode(RunMode runMode)
-        {
-            RunMode = runMode;
-            return this;
-        }
-        public CommandBuilder WithPriority(int priority)
-        {
-            Priority = priority;
-            return this;
-        }
 
         public CommandBuilder AddAliases(params string[] aliases)
         {
             _aliases.AddRange(aliases);
             return this;
         }
-        public CommandBuilder AddPrecondition(PreconditionAttribute precondition)
+        public CommandBuilder AddOverload(Action<OverloadBuilder> overloadBuilder)
         {
-            _preconditions.Add(precondition);
-            return this;
-        }
-        public CommandBuilder AddParameter(string name, Type type, Action<ParameterBuilder> createFunc)
-        {
-            var param = new ParameterBuilder(this, name, type);
-            createFunc(param);
-            _parameters.Add(param);
-            return this;
-        }
-        internal CommandBuilder AddParameter(Action<ParameterBuilder> createFunc)
-        {
-            var param = new ParameterBuilder(this);
-            createFunc(param);
-            _parameters.Add(param);
+            var overload = new OverloadBuilder(this);
+            overloadBuilder(overload);
+            _overloads.Add(overload);
             return this;
         }
 
@@ -100,19 +72,6 @@ namespace Discord.Commands.Builders
             //Default name to first alias
             if (Name == null)
                 Name = _aliases[0];
-
-            if (_parameters.Count > 0)
-            {
-                var lastParam = _parameters[_parameters.Count - 1];
-
-                var firstMultipleParam = _parameters.FirstOrDefault(x => x.IsMultiple);
-                if ((firstMultipleParam != null) && (firstMultipleParam != lastParam))
-                    throw new InvalidOperationException("Only the last parameter in a command may have the Multiple flag.");
-                
-                var firstRemainderParam = _parameters.FirstOrDefault(x => x.IsRemainder);
-                if ((firstRemainderParam != null) && (firstRemainderParam != lastParam))
-                    throw new InvalidOperationException("Only the last parameter in a command may have the Remainder flag.");
-            }
 
             return new CommandInfo(this, info, service);
         }
